Enable setup actions for setup phase only
